fix(cache-proposals): deduplicate metrics computation and fix non-atomic audit write
Extract shared similarity-window metrics logic into
similarity-metrics.utils.ts so the evaluator and dispatcher
stay in sync when formulas change.

Swap the write order in the outcome evaluator: append the audit
entry before updating proposal details, so a failed audit insert
does not permanently block retry via the outcome_evaluation guard.
