+++ proprietary/cache-proposals/cache-apply.dispatcher.ts
@@ -13,6 +13,7 @@ import { ConnectionRegistry } from '@app/connections/connection-registry.service
 import { CacheResolverService, type ResolvedCache } from './cache-resolver.service';
 import { ApplyFailedError } from './errors';
 import type { TuningMetricsSnapshot } from './cache-readonly.types';
+import { computeMetricsFromSimilarityWindow } from './similarity-metrics.utils';
 
 const FT_SEARCH_LIMIT = 1000;
 const TUNING_HISTORY_KEY_SUFFIX = ':__tuning_history';
@@ -177,65 +178,20 @@ export class CacheApplyDispatcher {
     prefix: string,
     threshold: number,
   ): Promise<{ signal: string; metrics: TuningMetricsSnapshot } | null> {
-    const DEFAULT_UNCERTAINTY_BAND = 0.05;
-    let raw: Array<string | number>;
-    try {
-      raw = (await client.zrange(
-        `${prefix}:__similarity_window`, '0', '-1', 'WITHSCORES',
-      )) as Array<string | number>;
-    } catch {
-      return null;
-    }
+    const result = await computeMetricsFromSimilarityWindow(client, prefix, threshold);
+    if (!result) return null;
 
-    // Read uncertainty_band from config (fall back to default).
-    let uncertaintyBand = DEFAULT_UNCERTAINTY_BAND;
-    try {
-      const configRaw = await client.hget(`${prefix}:__config`, 'uncertainty_band');
-      if (configRaw) {
-        const parsed = Number(configRaw);
-        if (Number.isFinite(parsed)) uncertaintyBand = parsed;
-      }
-    } catch { /* use default */ }
-
-    // Parse similarity window entries.
-    const hits: number[] = [];
-    const misses: number[] = [];
-    for (let i = 0; i < raw.length; i += 2) {
-      const member = raw[i];
-      if (typeof member !== 'string') continue;
-      try {
-        const entry = JSON.parse(member) as { score?: number; result?: string };
-        const score = typeof entry.score === 'number' ? entry.score : NaN;
-        if (!Number.isFinite(score)) continue;
-        if (entry.result === 'hit') hits.push(score);
-        else if (entry.result === 'miss') misses.push(score);
-      } catch { /* skip */ }
-    }
-
-    const total = hits.length + misses.length;
-    if (total === 0) return null;
-
-    const hitRate = hits.length / total;
-    const uncertainHitRate = hits.length === 0 ? 0
-      : hits.filter((s) => s >= threshold - uncertaintyBand).length / hits.length;
-    const midpoint = threshold / 2;
-    const distantHitRate = hits.length === 0 ? 0
-      : hits.filter((s) => s > midpoint).length / hits.length;
-    const nearMissRate = misses.length === 0 ? 0
-      : misses.filter((s) => s > threshold && s <= threshold + uncertaintyBand).length / misses.length;
+    const { metrics, hitCount, missCount } = result;
 
     // Determine dominant signal (must match the recommendation engine's guards exactly).
-    const uncertainFractionOfAll = uncertainHitRate * hitRate;
+    const uncertainFractionOfAll = metrics.uncertain_hit_rate * metrics.hit_rate;
     let signal = 'optimal';
-    if (uncertainHitRate > 0.2 && uncertainFractionOfAll > 0.15) signal = 'uncertain_hits';
-    else if (distantHitRate > 0.25 && hits.length >= 20 && hitRate > 0.8) signal = 'distant_hits';
-    else if (nearMissRate > 0.25) signal = 'near_misses';
-    else if (hitRate < 0.05 && misses.length >= 20) signal = 'low_hit_rate';
+    if (metrics.uncertain_hit_rate > 0.2 && uncertainFractionOfAll > 0.15) signal = 'uncertain_hits';
+    else if (metrics.distant_hit_rate > 0.25 && hitCount >= 20 && metrics.hit_rate > 0.8) signal = 'distant_hits';
+    else if (metrics.near_miss_rate > 0.25) signal = 'near_misses';
+    else if (metrics.hit_rate < 0.05 && missCount >= 20) signal = 'low_hit_rate';
 
-    return {
-      signal,
-      metrics: { hit_rate: hitRate, uncertain_hit_rate: uncertainHitRate, distant_hit_rate: distantHitRate, near_miss_rate: nearMissRate },
-    };
+    return { signal, metrics };
   }
 
   private async applyAgentToolTtlAdjust(
